feat(statestore): move adapter code and add migration
